Fix server installation
