+++ src/main/java/com/deku/cherryblossomgrotto/common/items/KabutoSandals.java
@@ -14,7 +14,7 @@ import org.jetbrains.annotations.Nullable;
 
 import java.util.function.Consumer;
 
-public class KabutoSandals extends ArmorItem implements IItemRenderProperties {
+public class KabutoSandals extends ArmorItem {
     public KabutoSandals() {
         super(ArmorMaterials.IRON, EquipmentSlot.FEET, new Properties().stacksTo(1).tab(CreativeModeTab.TAB_COMBAT));
         setRegistryName("kabuto_sandals");
