+++ src/main/java/com/deku/cherryblossomgrotto/common/items/NinjaLeggings.java
@@ -14,7 +14,7 @@ import org.jetbrains.annotations.Nullable;
 
 import java.util.function.Consumer;
 
-public class NinjaLeggings extends ArmorItem implements IItemRenderProperties {
+public class NinjaLeggings extends ArmorItem {
     public NinjaLeggings() {
         super(ModArmorMaterials.WOOL, EquipmentSlot.LEGS, new Item.Properties().stacksTo(1).tab(CreativeModeTab.TAB_COMBAT));
         setRegistryName("ninja_leggings");
